+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 
 * Fix parsing error during stats generation
 
+### Enhancements
+
+* Use `v2/submissions/` API endpoint to fetch programs
+* Handle program submissions with `delete` mode
+
 ## v25.11.9
 
 ### Enhancements